WIP fixing docs for 2.0
