--- docs/src/tutorials/intro_page.md
@@ -1,6 +1,0 @@
-# SIIP-Examples
-
-All the tutorials for the SIIP project are part of a separate repository
-[SIIP-Examples](https://github.com/NREL-SIIP/SIIPExamples.jl). You can access the latest
-`PowerSystems.jl` specific tutorial notebooks in
-[this link](https://nbviewer.jupyter.org/github/NREL-SIIP/SIIPExamples.jl/blob/master/notebook/2_PowerSystems_examples/).
